+++ app/Http/Controllers/UsersController.php
@@ -11,17 +11,12 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Hash;
 use Kami\Cocktail\Http\Requests\UserRequest;
 use Kami\Cocktail\Http\Resources\UserResource;
-use Kami\Cocktail\Search\SearchActionsAdapter;
 use Illuminate\Http\Resources\Json\JsonResource;
 
 class UsersController extends Controller
 {
     public function index(Request $request): JsonResource
     {
-        if (!$request->user()->isAdmin()) {
-            abort(403);
-        }
-
         $users = User::orderBy('id')->get();
 
         return UserResource::collection($users);
@@ -29,28 +24,18 @@ class UsersController extends Controller
 
     public function show(Request $request, int $id): JsonResource
     {
-        if (!$request->user()->isAdmin()) {
-            abort(403);
-        }
-
         $user = User::findOrFail($id);
 
         return new UserResource($user);
     }
 
-    public function store(SearchActionsAdapter $search, UserRequest $request): JsonResponse
+    public function store(UserRequest $request): JsonResponse
     {
-        if (!$request->user()->isAdmin()) {
-            abort(403);
-        }
-
         $user = new User();
         $user->name = $request->post('name');
         $user->email = $request->post('email');
         $user->email_verified_at = now();
         $user->password = Hash::make($request->post('password'));
-        $user->is_admin = (bool) $request->post('is_admin');
-        $user->search_api_key = $search->getActions()->getPublicApiKey();
         $user->save();
 
         return (new UserResource($user))
@@ -61,15 +46,10 @@ class UsersController extends Controller
 
     public function update(int $id, UserRequest $request): JsonResource
     {
-        if (!$request->user()->isAdmin() || $id === 1) {
-            abort(403);
-        }
-
         $user = User::findOrFail($id);
         $user->name = $request->post('name');
         $user->email = $request->post('email');
         $user->email_verified_at = now();
-        $user->is_admin = (bool) $request->post('is_admin');
 
         if ($request->has('password')) {
             $user->password = Hash::make($request->post('password'));
@@ -82,10 +62,6 @@ class UsersController extends Controller
 
     public function delete(Request $request, int $id): Response
     {
-        if (!$request->user()->isAdmin() || $id === 1) {
-            abort(403);
-        }
-
         User::findOrFail($id)->delete();
 
         return response(null, 204);