+++ app/Http/Resources/BarResource.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kami\Cocktail\Http\Resources;
+
+use Illuminate\Http\Resources\Json\JsonResource;
+
+/**
+ * @mixin \Kami\Cocktail\Models\Bar
+ */
+class BarResource extends JsonResource
+{
+    /**
+     * Transform the resource into an array.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return array<string, mixed>
+     */
+    public function toArray($request)
+    {
+        return [
+            'id' => $this->id,
+            'name' => $this->name,
+            'subtitle' => $this->subtitle,
+            'description' => $this->description,
+            'search_driver_api_key' => $this->search_driver_api_key,
+        ];
+    }
+}