+++ routes/api.php
@@ -84,8 +84,8 @@ Route::middleware($authMiddleware)->group(function() {
     });
 
     Route::prefix('ingredient-categories')->group(function() {
-        Route::get('/', [IngredientCategoryController::class, 'index']);
-        Route::post('/', [IngredientCategoryController::class, 'store']);
+        Route::get('/', [IngredientCategoryController::class, 'index'])->middleware(EnsureRequestHasBarQuery::class);
+        Route::post('/', [IngredientCategoryController::class, 'store'])->middleware(EnsureRequestHasBarQuery::class);
         Route::get('/{id}', [IngredientCategoryController::class, 'show'])->name('ingredient-categories.show');
         Route::put('/{id}', [IngredientCategoryController::class, 'update']);
         Route::delete('/{id}', [IngredientCategoryController::class, 'delete']);